add complex64s

+++ safe/float64s.go
@@ -12,11 +12,11 @@ func (s Float64s) Slice(start, stop int) Float64s {
 	return Float64s{s.slice.slice(start, stop, sizeofFloat64)}
 }
 
-func (dst Float64s) CopyHtoD(src []float32) {
+func (dst Float64s) CopyHtoD(src []float64) {
 	dst.copyHtoD(unsafe.Pointer(&src[0]), len(src), sizeofFloat64)
 }
 
-func (src Float64s) CopyDtoH(dst []float32) {
+func (src Float64s) CopyDtoH(dst []float64) {
 	src.copyDtoH(unsafe.Pointer(&dst[0]), len(dst), sizeofFloat64)
 }
